# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Decent Event Management API",
        "description": "Event ticketing and management system API with registration, payment, ticketing, check-in, and reporting capabilities.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.example.com/v1",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/admin/auth/2fa/setup": {
            "post": {
                "tags": [
                    "Two-Factor"
                ],
                "summary": "Generate a new TOTP secret and QR code for the authenticated staff member",
                "description": "Self-service: acts on the caller's own account only, no RBAC permission check beyond authentication. Available to a token holding either the `admin` or the setup-only `2fa-setup` ability.",
                "operationId": "0e03408bb7659e657a48352d96004c4c",
                "responses": {
                    "200": {
                        "description": "TOTP secret generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "secret": {
                                            "description": "TOTP secret key",
                                            "type": "string"
                                        },
                                        "qr_code_svg": {
                                            "description": "QR code as inline SVG markup",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "2FA is already enabled — disable it first to reconfigure"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/auth/2fa/confirm": {
            "post": {
                "tags": [
                    "Two-Factor"
                ],
                "summary": "Confirm 2FA setup with a TOTP code and activate it",
                "description": "Self-service: acts on the caller's own account only, no RBAC permission check beyond authentication. Requires a prior call to /admin/auth/2fa/setup. On success, revokes the setup-only token and issues a full `admin`-ability token.",
                "operationId": "78af7daee660c6ee45fda8686134faa9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "description": "6-digit TOTP code",
                                        "required": [
                                            "code"
                                        ],
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA enabled; a new full-access token is issued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "recovery_codes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "token": {
                                            "description": "New full-access API bearer token",
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing/invalid code, or setup was never called"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/auth/2fa/disable": {
            "post": {
                "tags": [
                    "Two-Factor"
                ],
                "summary": "Disable 2FA for the authenticated staff member",
                "description": "Self-service: acts on the caller's own account only, no RBAC permission check beyond authentication. Requires the caller to re-enter their password. Only reachable with a full `admin`-ability token, i.e. 2FA must already be confirmed.",
                "operationId": "a68367b94235a211d1e74c15813f8770",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "required": [
                                            "password"
                                        ],
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing or incorrect password"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token issued by the relevant login endpoint. Admin, attendee, and scanner guards each mint tokens scoped to that guard and cannot be used interchangeably.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}